Updates based on CR feedback
