+++ tests/logging/pipeline_logger_test.c
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-#include <aws/io/pipeline_logger.h>
+#include <aws/io/logging.h>
 
 #include <aws/testing/aws_test_harness.h>
 
@@ -36,7 +36,7 @@ int do_pipeline_logger_test(log_test_fn log_fn, const char **expected_user_conte
     remove(s_test_file_name);
 
     struct aws_logger logger;
-    if (aws_pipeline_logger_file_writer_init(&logger, aws_default_allocator(), AWS_LL_TRACE, s_test_file_name)) {
+    if (aws_logger_standard_init(&logger, aws_default_allocator(), AWS_LL_TRACE, s_test_file_name)) {
         return AWS_OP_ERR;
     }
